+++ app/src/main/java/com/concordium/wallet/ui/bakerdelegation/common/segmentedview/SegmentedButtonView.kt
@@ -19,7 +19,12 @@ class SegmentedButtonView @JvmOverloads constructor(
         binding = SegmentedTextItemBinding.bind(this)
     }
 
-    override fun setLayout(title: String, earningPercent: String, selected: Boolean) {
+    override fun setLayout(
+        title: String,
+        description: String,
+        earningPercent: String,
+        selected: Boolean
+    ) {
         binding.button.text = title
         binding.button.isSelected = selected
     }