+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] 
+
+### Fixed 
+
+- Delegation Register pool screen UI bugs
+
+## [1.15.1] - 2025-11-06
 
 ### Fixed
 
 - The Send screen Max button not working properly with large balances
+
+### Changed
+
+- Updated staking flow texts
 
 ## [1.15.0] - 2025-10-27
 
@@ -324,7 +334,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   attributes
 - CIS-2 tokens with corrupted or missing metadata can no longer be added
 
-[Unreleased]: https://github.com/Concordium/cryptox-android/compare/1.15.0...HEAD
+[Unreleased]: https://github.com/Concordium/cryptox-android/compare/1.15.1...HEAD
+
+[1.15.1]: https://github.com/Concordium/cryptox-android/compare/1.15.0...1.15.1
 
 [1.15.0]: https://github.com/Concordium/cryptox-android/compare/1.14.0...1.15.0